Fixed remider services
git-svn-id: https://svn.iportalen.dk/svn-private/F2007/trunk@91 d9c49cb3-4128-44d0-8390-2fdd9ac6ca64

+++ core/src/main/java/dk/bregnvig/formula1/event/RaceTimer.java
@@ -21,5 +21,5 @@ public interface RaceTimer {
 	 * @param race
 	 * @return
 	 */
-	int getDelay(Race race);
+	long getDelay(Race race);
 }
